requirements.txt
flask
requests
python-dotenv
websocket-client
cloudscraper
keyboard